Update Full folders api

+++ src/controller/FileController.ts
@@ -1,13 +1,14 @@
-import {getRepository, Index, createQueryBuilder} from "typeorm";
+import {getRepository, Index, createQueryBuilder, getTreeRepository , getManager,getConnection} from "typeorm";
 import {NextFunction, Request, Response} from "express";
 import {Folder} from "../entity/Folder";
 import * as fs from "fs";
 import * as path from "path";
-import {checkRenameFile, checkFilterFile,checkLimit,checkOffset} from '../checks/Validate';
+import {checkRenameFile, checkFilterFile,checkLimit,checkOffset,checkFoldersSchema, checkType} from '../checks/Validate';
 import { File } from "../entity/File";
 const Joi = require('joi');
 var multer = require('multer');
 import {config} from "../../config/config";
+import { ClientResponse } from "http";
 
 
 
@@ -15,9 +16,9 @@ export class FileController {
 
     private fileRepository = getRepository(File);
     private folderRepository = getRepository(Folder);
+    private folderTreeRepository = getTreeRepository(Folder);
 
-    async createFile(request: Request, response: Response, next: NextFunction){
-
+    async createFile(request:Request, response: Response ){
         const url = path.posix.join(__dirname,config.storage);
         const self = this;
         const folder = await this.folderRepository.findOne({id: request.params.foldersId});
@@ -33,9 +34,10 @@ export class FileController {
 
         upload(request,response,function(err) {
             if(err) {
-                return response.status(403).json({
-                            message: "Error uploading file"
-                        });
+                response.status(403).json({
+                    message: "Error uploading file"
+                });
+                return;
             } else {
                 const files = folder.files;
                 for (let i = 0; i < files.length; i++){
@@ -64,123 +66,47 @@ export class FileController {
             }
         });
 
-
     }
 
-    async readFiles(request: Request, response: Response, next: NextFunction){
-        let offsetObj = {offset: 0};
-        let limitObj = {limit: 0};
-        let filterObj = {type: ''};
-        offsetObj.offset = Number(request.query.offset);
-        limitObj.limit = Number(request.query.limit);
-        filterObj.type = request.query.type;
-        if(!checkLimit(limitObj,response)){
-            request.query.limit = undefined;
-        }
-        if(!checkOffset(offsetObj,response)){
-            request.query.offset = undefined;
-        }
-        if((request.query.offset == undefined) && ( request.query.limit == undefined)){
-            if(!checkFilterFile(filterObj,response)){
-                const folder = await this.folderRepository.findOne({id : request.params.foldersId});
-                if( folder !== undefined){
-                    if( folder.files.length !== 0){
-                        return folder.files;
-                    } else {
-                        response.status(404).json({
-                            message: 'Files not found'
-                        });
-                        return;
-                    }
-                } else {
-                    response.status(404).json({
-                        message: 'Folder not found'
-                    });
-                    return;
-                }
-            };
-            const files = await this.fileRepository.find({folder: {id: request.params.foldersId},type: request.query.type});
-            if( files.length === 0){
-                response.status(404).json({
-                    message: 'Files with this type doesn\'t exist'
-                });
-                return;
-            }
-            return files;
-        } else {
-            if(request.query.limit == undefined){
-                request.query.offset = Number(request.query.offset);
-                const files =  await this.fileRepository.find({relations: ["folder"],where:{folder:{id: request.params.foldersId}},skip: request.query.offset});
-                if( files.length === 0){
-                    response.status(404).json({
-                        message: 'Files doesn\'t exist'
-                    });
-                    return;
-                }
-                let [{id, fileName, status, type, date}] = files;
-                return [{id, fileName, status, type, date}];
-            }
-            if(request.query.offset == undefined){
-                request.query.limit = Number(request.query.limit);
-                const files =  await this.fileRepository.find({relations: ["folder"],where:{folder:{id: request.params.foldersId}},take: request.query.limit});
-                if( (files.length === 0)){
-                    response.status(404).json({
-                        message: 'Files doesn\'t exist'
-                    });
-                    return;
-                };
-                return files.map(({id, fileName, status, type, date}) => ({id, fileName, status, type, date}));
-            };
-            request.query.offset = Number(request.query.offset);
-            request.query.limit = Number(request.query.limit);
-            const files =  await this.fileRepository.find({relations: ["folder"],where:{folder:{id: request.params.foldersId}},skip: request.query.offset,take: request.query.limit});
-            if( files.length === 0){
-                response.status(404).json({
-                    message: 'Files doesn\'t exist'
-                });
-                return;
-            };
-            return files.map(({id, fileName, status, type, date}) => ({id, fileName, status, type, date}));
-        };
-
-
-    }
-
-    async readFile(request: Request, response: Response, next: NextFunction){
-        const folder = await this.folderRepository.findOne({id : request.params.foldersId});
-        const file = await this.fileRepository.findOne({id : request.params.filesId});
-        if( folder !== undefined ){
-            if( folder.files.length !== 0){
-                let files = folder.files;
-                if(file !== undefined){
-                    for( let i = 0; i < files.length; i++){
-                        if(files[i].id === file.id){
-                            return folder.files[i];
-                        }
-                    }
-                } else {
-                    response.status(404).json({
-                        message: 'File not found'
-                    });
-                    return;
-                }
-            } else {
-                response.status(404).json({
-                    message: 'Files doesn\'t exist'
-                });
-                return;
-            }
-        } else {
-            response.status(404).json({
-                message:' Folder not found'
-            });
+    async createFolder(request: Request, response: Response){
+        let folderObj = {folderName: ''};
+        folderObj.folderName = request.body.folderName;
+        if(!checkFoldersSchema(folderObj,response)){
             return;
-        }
+        };
+        let folderParent = await this.folderRepository.findOne({id : request.params.foldersId}); 
+        const folder = new Folder();
+        folder.folderName = request.body.folderName;
+        folder.status = "Created";
+        folder.date = Date.now().toLocaleString();
+        folder.parentFolder = folderParent;
+        let {id, folderName, status, date} = folder;
+        getManager().save(folder);
+        response.status(201).json({
+            message: 'Folder successfuly created',
+            folder: {id, folderName, status, date}
+        })
+        return;
     }
 
-    async deleteFile(request: Request, response: Response, next: NextFunction){
+    async createResource(request: Request, response: Response, next: NextFunction){
+        let typeEntity = {type : ''};
+        if(request.body.length !== undefined){
+            typeEntity.type = request.body.type;
+            if(!checkType(typeEntity,response)){
+                response.status(401).json({
+                    message: "Invalid request data"
+                });
+                return;
+            };
+        }
+        request.body.type === "folder" ?  this.createFolder(request,response) : this.createFile(request,response);
+        return;
+    }
+
+    async deleteFile(request: Request, response: Response){
         const folder = await this.folderRepository.findOne({id: request.params.foldersId});
-        const file = await this.fileRepository.findOne({id : request.params.filesId});
+        const file = await this.fileRepository.findOne({id : request.params.resourcesId});
         const self = this;
         if( folder !== undefined ){
             if( folder.files.length !== 0){
@@ -224,38 +150,40 @@ export class FileController {
         }
     }
 
+    async deleteFolder(child: Folder, response: Response){
+        console.log(child);
+        await this.folderRepository.delete({id: child.id}).then((res) => console.log(res)).catch(err => console.log("Error lala", err));
+    }
 
-    async deleteFiles(request: Request, response: Response, next: NextFunction){
-        const folder = await this.folderRepository.findOne({id: request.params.foldersId});
-        const file = await this.fileRepository.findOne({id : request.params.filesId});
-        const self = this;
-        if( folder !== undefined ){
-            if( folder.files.length !== 0){
-                for(let i = 0; i < folder.files.length; i++){
-                    const url = path.posix.join(__dirname, config.storage + folder.files[i].fileName.slice(0,folder.files[i].fileName.lastIndexOf('.')) + folder.id + path.extname(folder.files[i].fileName));
-                    fs.stat(url, (err, stats) => {
-                        if(err)
-                            return err;
-                        fs.unlink(url, (err) => {
-                            if(err)
-                                return err;
-                            self.fileRepository.remove(folder.files[i]);    
-                        })
-                    })
-                }
-                response.status(200).json({
-                    message: 'File\'s deleted'
-                });
-                return;  
-            } else {
+
+    async deleteResource(request: Request, response: Response, next: NextFunction){
+        let arrFolderChildren = [];
+        let folder = await this.folderRepository.findOne({id : request.params.foldersId});
+        if( folder !== undefined){
+            let children = await this.folderTreeRepository.findDescendantsTree(folder);
+            if( (children.files.length === 0) && (children.childFolders.length === 0)){
                 response.status(404).json({
-                    message: 'Files doesn\'t exist'
+                    message: 'Children don\'t exsits'
                 });
                 return;
+            };
+            const childFoldersArray = children.childFolders.map(({id, folderName, status, date, files, childFolders}) => ({id, folderName, status, date, files, childFolders}));
+            arrFolderChildren = arrFolderChildren.concat(childFoldersArray);
+            arrFolderChildren = arrFolderChildren.concat(children.files.map(({id, fileName, status,type, date}) => ({id, fileName, status,type , date})));
+            console.log(arrFolderChildren);
+            for (let child of arrFolderChildren){
+                if(child.id == request.params.resourcesId){
+                    return !child.type ? await this.deleteFolder(child,response) : this.deleteFile(request,response);
+                }
             }
+            response.status(404).json({
+                message: "Resource doesn\'t exist"
+            });
+            return;
+
         } else {
             response.status(404).json({
-                message:' Folder not found'
+                message: 'Folder not found'
             });
             return;
         }
@@ -263,8 +191,9 @@ export class FileController {
 
     async renameFile(request: Request, response: Response, next: NextFunction){
         const folder = await this.folderRepository.findOne({id : request.params.foldersId});
-        const file = await this.fileRepository.findOne({id : request.params.filesId});
+        const file = await this.fileRepository.findOne({id: request.params.resourcesId});
         let files = folder.files;
+        let flagExistFile = false;
         if(!checkRenameFile(request.body,response)){
             return;
         }
@@ -275,32 +204,44 @@ export class FileController {
             return;
         }
         for(let i = 0; i < files.length; i++){
-            if(files[i].fileName === (request.body.fileName + path.extname(file.fileName))){
-                response.status(403).json({
-                    message: 'File with this name exist'
-                })
-                return;
+            if( files[i].id == request.params.resourcesId ){
+                flagExistFile = true;
             }
         }
-        for(let i = 0 ; i < files.length; i++){
-            if(files[i].fileName === file.fileName){
-                const url = path.posix.join(__dirname, config.storage + files[i].fileName.slice(0,files[i].fileName.lastIndexOf('.')) + folder.id + path.extname(files[i].fileName));
-                fs.stat(url, (err, stats) => {
-                    if(err)
-                        return err;
-                    const newUrl = path.posix.join(__dirname,config.storage + request.body.fileName + folder.id + path.extname(files[i].fileName));
-                    fs.rename(url, newUrl,(err) => {
+        if(flagExistFile){
+            for(let i = 0; i < files.length; i++){
+                if(files[i].fileName === (request.body.fileName + path.extname(file.fileName))){
+                    response.status(403).json({
+                        message: 'File with this name exist'
+                    })
+                    return;
+                }
+            }
+            for(let i = 0 ; i < files.length; i++){
+                if(files[i].fileName === file.fileName){
+                    const url = path.posix.join(__dirname, config.storage + files[i].fileName.slice(0,files[i].fileName.lastIndexOf('.')) + folder.id + path.extname(files[i].fileName));
+                    fs.stat(url, (err, stats) => {
                         if(err)
                             return err;
-                        files[i].fileName = request.body.fileName + path.extname(files[i].fileName);
-                        this.fileRepository.save(files[i]);
-                        response.status(200).json({
-                            message: "File renamed"
+                        const newUrl = path.posix.join(__dirname,config.storage + request.body.fileName + folder.id + path.extname(files[i].fileName));
+                        fs.rename(url, newUrl,(err) => {
+                            if(err)
+                                return err;
+                            files[i].fileName = request.body.fileName + path.extname(files[i].fileName);
+                            this.fileRepository.save(files[i]);
+                            response.status(200).json({
+                                message: "File renamed"
+                            });
+                            return;    
                         });
-                        return;    
-                    });
-                })
+                    })
+                }
             }
+        } else {
+            response.json({
+                message: 'File not found'
+            });
+            return;
         }
     }
